requirements.txt
certifi==2024.8.30
charset-normalizer==3.3.2
icalendar==6.0.0
idna==3.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
six==1.16.0
todoist_api_python==2.1.7
tzdata==2024.2
urllib3==2.2.3
